russh_sftp::client
Struct SftpSession

pub struct SftpSession { /* private fields */ }

High-level SFTP implementation for easy interaction with a remote file system. Contains most methods similar to the native filesystem

Implementations

impl SftpSession

pub async fn new<S>(stream: S) -> SftpResult<Self>
where S: AsyncRead + AsyncWrite + Unpin + Send + 'static,

Creates a new session by initializing the protocol and extensions

pub async fn new_opts<S>(stream: S, timeout: Option<u64>) -> SftpResult<Self>
where S: AsyncRead + AsyncWrite + Unpin + Send + 'static,

Creates a new session with timeout opt before the first request

pub async fn set_timeout(&self, secs: u64)

Set the maximum response time in seconds. Default: 10 seconds

pub async fn close(&self) -> SftpResult<()>

Closes the inner channel stream.

pub async fn open<T: Into<String>>(&self, filename: T) -> SftpResult<File>

Attempts to open a file in read-only mode.

pub async fn create<T: Into<String>>(&self, filename: T) -> SftpResult<File>

Opens a file in write-only mode.

This function will create a file if it does not exist, and will truncate it if it does.

pub async fn open_with_flags<T: Into<String>>( &self, filename: T, flags: OpenFlags, ) -> SftpResult<File>

Attempts to open or create the file in the specified mode

pub async fn open_with_flags_and_attributes<T: Into<String>>( &self, filename: T, flags: OpenFlags, attributes: FileAttributes, ) -> SftpResult<File>

Attempts to open or create the file in the specified mode and with specified file attributes

pub async fn canonicalize<T: Into<String>>(&self, path: T) -> SftpResult<String>

Requests the remote party for the absolute from the relative path.

pub async fn create_dir<T: Into<String>>(&self, path: T) -> SftpResult<()>

Creates a new empty directory.

pub async fn read<P: Into<String>>(&self, path: P) -> SftpResult<Vec<u8>>

Reads the contents of a file located at the specified path to the end.

pub async fn write<P: Into<String>>( &self, path: P, data: &[u8], ) -> SftpResult<()>

Writes the contents to a file whose path is specified.

pub async fn try_exists<P: Into<String>>(&self, path: P) -> SftpResult<bool>

Checks a file or folder exists at the specified path

pub async fn read_dir<P: Into<String>>(&self, path: P) -> SftpResult<ReadDir>

Returns an iterator over the entries within a directory.

pub async fn read_link<P: Into<String>>(&self, path: P) -> SftpResult<String>

Reads a symbolic link, returning the file that the link points to.

pub async fn remove_dir<P: Into<String>>(&self, path: P) -> SftpResult<()>

Removes the specified folder.

pub async fn remove_file<T: Into<String>>(&self, filename: T) -> SftpResult<()>

Removes the specified file.

pub async fn rename<O, N>(&self, oldpath: O, newpath: N) -> SftpResult<()>
where O: Into<String>, N: Into<String>,

Rename a file or directory to a new name.

pub async fn symlink<P, T>(&self, path: P, target: T) -> SftpResult<()>
where P: Into<String>, T: Into<String>,

Creates a symlink of the specified target.

pub async fn metadata<P: Into<String>>(&self, path: P) -> SftpResult<Metadata>

Queries metadata about the remote file.

pub async fn set_metadata<P: Into<String>>( &self, path: P, metadata: Metadata, ) -> Result<(), Error>

Sets metadata for a remote file.

pub async fn symlink_metadata<P: Into<String>>( &self, path: P, ) -> SftpResult<Metadata>

pub async fn hardlink<O, N>(&self, oldpath: O, newpath: N) -> SftpResult<bool>
where O: Into<String>, N: Into<String>,

pub async fn fs_info<P: Into<String>>( &self, path: P, ) -> SftpResult<Option<Statvfs>>

Performs a statvfs on the remote file system path. Returns [Ok(None)] if the remote SFTP server does not support statvfs@openssh.com extension v2.