russh_sftp::server
Trait Handler

pub trait Handler: Sized {
    type Error: Into<StatusCode>;

    // Required method
    fn unimplemented(&self) -> Self::Error;

    // Provided methods
    fn init<'life0, 'async_trait>(
        &'life0 mut self,
        version: u32,
        extensions: HashMap<String, String>,
    ) -> Pin<Box<dyn Future<Output = Result<Version, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
    fn open<'life0, 'async_trait>(
        &'life0 mut self,
        id: u32,
        filename: String,
        pflags: OpenFlags,
        attrs: FileAttributes,
    ) -> Pin<Box<dyn Future<Output = Result<Handle, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
    fn close<'life0, 'async_trait>(
        &'life0 mut self,
        id: u32,
        handle: String,
    ) -> Pin<Box<dyn Future<Output = Result<Status, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
    fn read<'life0, 'async_trait>(
        &'life0 mut self,
        id: u32,
        handle: String,
        offset: u64,
        len: u32,
    ) -> Pin<Box<dyn Future<Output = Result<Data, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
    fn write<'life0, 'async_trait>(
        &'life0 mut self,
        id: u32,
        handle: String,
        offset: u64,
        data: Vec<u8>,
    ) -> Pin<Box<dyn Future<Output = Result<Status, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
    fn lstat<'life0, 'async_trait>(
        &'life0 mut self,
        id: u32,
        path: String,
    ) -> Pin<Box<dyn Future<Output = Result<Attrs, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
    fn fstat<'life0, 'async_trait>(
        &'life0 mut self,
        id: u32,
        handle: String,
    ) -> Pin<Box<dyn Future<Output = Result<Attrs, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
    fn setstat<'life0, 'async_trait>(
        &'life0 mut self,
        id: u32,
        path: String,
        attrs: FileAttributes,
    ) -> Pin<Box<dyn Future<Output = Result<Status, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
    fn fsetstat<'life0, 'async_trait>(
        &'life0 mut self,
        id: u32,
        handle: String,
        attrs: FileAttributes,
    ) -> Pin<Box<dyn Future<Output = Result<Status, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
    fn opendir<'life0, 'async_trait>(
        &'life0 mut self,
        id: u32,
        path: String,
    ) -> Pin<Box<dyn Future<Output = Result<Handle, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
    fn readdir<'life0, 'async_trait>(
        &'life0 mut self,
        id: u32,
        handle: String,
    ) -> Pin<Box<dyn Future<Output = Result<Name, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
    fn remove<'life0, 'async_trait>(
        &'life0 mut self,
        id: u32,
        filename: String,
    ) -> Pin<Box<dyn Future<Output = Result<Status, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
    fn mkdir<'life0, 'async_trait>(
        &'life0 mut self,
        id: u32,
        path: String,
        attrs: FileAttributes,
    ) -> Pin<Box<dyn Future<Output = Result<Status, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
    fn rmdir<'life0, 'async_trait>(
        &'life0 mut self,
        id: u32,
        path: String,
    ) -> Pin<Box<dyn Future<Output = Result<Status, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
    fn realpath<'life0, 'async_trait>(
        &'life0 mut self,
        id: u32,
        path: String,
    ) -> Pin<Box<dyn Future<Output = Result<Name, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
    fn stat<'life0, 'async_trait>(
        &'life0 mut self,
        id: u32,
        path: String,
    ) -> Pin<Box<dyn Future<Output = Result<Attrs, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
    fn rename<'life0, 'async_trait>(
        &'life0 mut self,
        id: u32,
        oldpath: String,
        newpath: String,
    ) -> Pin<Box<dyn Future<Output = Result<Status, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
    fn readlink<'life0, 'async_trait>(
        &'life0 mut self,
        id: u32,
        path: String,
    ) -> Pin<Box<dyn Future<Output = Result<Name, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
    fn symlink<'life0, 'async_trait>(
        &'life0 mut self,
        id: u32,
        linkpath: String,
        targetpath: String,
    ) -> Pin<Box<dyn Future<Output = Result<Status, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
    fn extended<'life0, 'async_trait>(
        &'life0 mut self,
        id: u32,
        request: String,
        data: Vec<u8>,
    ) -> Pin<Box<dyn Future<Output = Result<Packet, Self::Error>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait { ... }

}

Server handler for each client. This is async_trait

Required Associated Types

type Error: Into<StatusCode>

The type must have an Into<StatusCode> implementation because a response must be sent to any request, even if completed by error.

Required Methods

fn unimplemented(&self) -> Self::Error

Called by the handler when the packet is not implemented

Provided Methods

fn init<'life0, 'async_trait>( &'life0 mut self, version: u32, extensions: HashMap<String, String>, ) -> Pin<Box<dyn Future<Output = Result<Version, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait,

The default is to send an SSH_FXP_VERSION response with the protocol version and ignore any extensions.

fn open<'life0, 'async_trait>( &'life0 mut self, id: u32, filename: String, pflags: OpenFlags, attrs: FileAttributes, ) -> Pin<Box<dyn Future<Output = Result<Handle, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait,

Called on SSH_FXP_OPEN

fn close<'life0, 'async_trait>( &'life0 mut self, id: u32, handle: String, ) -> Pin<Box<dyn Future<Output = Result<Status, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait,

Called on SSH_FXP_CLOSE. The status can be returned as Ok or as Err

fn read<'life0, 'async_trait>( &'life0 mut self, id: u32, handle: String, offset: u64, len: u32, ) -> Pin<Box<dyn Future<Output = Result<Data, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait,

Called on SSH_FXP_READ

fn write<'life0, 'async_trait>( &'life0 mut self, id: u32, handle: String, offset: u64, data: Vec<u8>, ) -> Pin<Box<dyn Future<Output = Result<Status, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait,

Called on SSH_FXP_WRITE

fn lstat<'life0, 'async_trait>( &'life0 mut self, id: u32, path: String, ) -> Pin<Box<dyn Future<Output = Result<Attrs, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait,

Called on SSH_FXP_LSTAT

fn fstat<'life0, 'async_trait>( &'life0 mut self, id: u32, handle: String, ) -> Pin<Box<dyn Future<Output = Result<Attrs, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait,

Called on SSH_FXP_FSTAT

fn setstat<'life0, 'async_trait>( &'life0 mut self, id: u32, path: String, attrs: FileAttributes, ) -> Pin<Box<dyn Future<Output = Result<Status, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait,

Called on SSH_FXP_SETSTAT

fn fsetstat<'life0, 'async_trait>( &'life0 mut self, id: u32, handle: String, attrs: FileAttributes, ) -> Pin<Box<dyn Future<Output = Result<Status, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait,

Called on SSH_FXP_FSETSTAT

fn opendir<'life0, 'async_trait>( &'life0 mut self, id: u32, path: String, ) -> Pin<Box<dyn Future<Output = Result<Handle, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait,

Called on SSH_FXP_OPENDIR

fn readdir<'life0, 'async_trait>( &'life0 mut self, id: u32, handle: String, ) -> Pin<Box<dyn Future<Output = Result<Name, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait,

Called on SSH_FXP_READDIR. EOF error should be returned at the end of reading the directory

fn remove<'life0, 'async_trait>( &'life0 mut self, id: u32, filename: String, ) -> Pin<Box<dyn Future<Output = Result<Status, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait,

Called on SSH_FXP_REMOVE. The status can be returned as Ok or as Err

fn mkdir<'life0, 'async_trait>( &'life0 mut self, id: u32, path: String, attrs: FileAttributes, ) -> Pin<Box<dyn Future<Output = Result<Status, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait,

Called on SSH_FXP_MKDIR

fn rmdir<'life0, 'async_trait>( &'life0 mut self, id: u32, path: String, ) -> Pin<Box<dyn Future<Output = Result<Status, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait,

Called on SSH_FXP_RMDIR. The status can be returned as Ok or as Err

fn realpath<'life0, 'async_trait>( &'life0 mut self, id: u32, path: String, ) -> Pin<Box<dyn Future<Output = Result<Name, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait,

Called on SSH_FXP_REALPATH. Must contain only one name and a dummy attributes

fn stat<'life0, 'async_trait>( &'life0 mut self, id: u32, path: String, ) -> Pin<Box<dyn Future<Output = Result<Attrs, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait,

Called on SSH_FXP_STAT

fn rename<'life0, 'async_trait>( &'life0 mut self, id: u32, oldpath: String, newpath: String, ) -> Pin<Box<dyn Future<Output = Result<Status, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait,

Called on SSH_FXP_RENAME. The status can be returned as Ok or as Err

fn readlink<'life0, 'async_trait>( &'life0 mut self, id: u32, path: String, ) -> Pin<Box<dyn Future<Output = Result<Name, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait,

Called on SSH_FXP_READLINK

fn symlink<'life0, 'async_trait>( &'life0 mut self, id: u32, linkpath: String, targetpath: String, ) -> Pin<Box<dyn Future<Output = Result<Status, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait,

Called on SSH_FXP_SYMLINK. The status can be returned as Ok or as Err

fn extended<'life0, 'async_trait>( &'life0 mut self, id: u32, request: String, data: Vec<u8>, ) -> Pin<Box<dyn Future<Output = Result<Packet, Self::Error>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait,

Called on SSH_FXP_EXTENDED. The extension can return any packet, so itโs not specific. If the server does not recognize the `requestโ name the server must respond with an SSH_FX_OP_UNSUPPORTED error

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.